Plot improvements and adds variable column(s) filter
- plots uses as much as possible names and units for all axes
- you can now use:
	- single_col=var['column']
	- multi_col=var['col1', 'col2']
	- multi_col=var[['col1', 'col2']]
